+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - resolve changelog update regex and restore CHANGELOG.md
+
 ## [2.22.0] - 2025-12-07
 
 ### Added